(#7788) Allow gem directories to be searched while autoloading
This allows rubygems to be used to extend or override puppet behavior.

Unfortunately the features system uses the autoloader so checking for
the rubygems feature causes a call loop.  It's possible that a more
elegant way to avoid this exist rather than just checking the @path
but this did what I needed to do.

Without caching the directories the tests would take forever to run
and eventually crash my server.  Caching seems to avoid the biggest
of the performance hit.

Reviewed-By: Daniel Pittman <daniel@puppetlabs.com>